refactor(utils): move another function to utils
