+++ apps/backend/src/app_utils.rs
@@ -1,26 +1,17 @@
 use std::sync::Arc;
 
-use apalis::prelude::{MemoryStorage, MessageQueue};
-use async_graphql::{extensions::Tracing, EmptySubscription, MergedObject, Result, Schema};
+use apalis::prelude::MemoryStorage;
+use async_graphql::{extensions::Tracing, EmptySubscription, MergedObject, Schema};
 use background::{ApplicationJob, CoreApplicationJob};
-use chrono::Utc;
-use models::{
-    collection, collection_to_entity,
-    functions::associate_user_with_entity,
-    prelude::{Collection, CollectionToEntity, UserToCollection},
-    user_to_collection, ChangeCollectionToEntityInput,
-};
 use openidconnect::{
     core::{CoreClient, CoreProviderMetadata},
     reqwest::async_http_client,
     ClientId, ClientSecret, IssuerUrl, RedirectUrl,
 };
 use resolvers::{ExporterMutation, ExporterQuery};
-use sea_orm::{
-    ActiveModelTrait, ActiveValue, ColumnTrait, DatabaseConnection, EntityTrait, QueryFilter,
-};
+use sea_orm::DatabaseConnection;
 use services::{ExporterService, FileStorageService};
-use utils::{CteColAlias, FRONTEND_OAUTH_ENDPOINT};
+use utils::FRONTEND_OAUTH_ENDPOINT;
 
 use crate::{
     fitness::resolver::{ExerciseMutation, ExerciseQuery, ExerciseService},
@@ -121,77 +112,6 @@ pub async fn create_app_services(
         exporter_service,
         exercise_service,
     }
-}
-
-// TODO: Move this to database utils crate
-pub async fn add_entity_to_collection(
-    db: &DatabaseConnection,
-    user_id: &String,
-    input: ChangeCollectionToEntityInput,
-    perform_core_application_job: &MemoryStorage<CoreApplicationJob>,
-) -> Result<bool> {
-    let collection = Collection::find()
-        .left_join(UserToCollection)
-        .filter(user_to_collection::Column::UserId.eq(user_id))
-        .filter(collection::Column::Name.eq(input.collection_name))
-        .one(db)
-        .await
-        .unwrap()
-        .unwrap();
-    let mut updated: collection::ActiveModel = collection.into();
-    updated.last_updated_on = ActiveValue::Set(Utc::now());
-    let collection = updated.update(db).await.unwrap();
-    let resp = if let Some(etc) = CollectionToEntity::find()
-        .filter(CteColAlias::CollectionId.eq(collection.id.clone()))
-        .filter(
-            CteColAlias::MetadataId
-                .eq(input.metadata_id.clone())
-                .or(CteColAlias::PersonId.eq(input.person_id.clone()))
-                .or(CteColAlias::MetadataGroupId.eq(input.metadata_group_id.clone()))
-                .or(CteColAlias::ExerciseId.eq(input.exercise_id.clone()))
-                .or(CteColAlias::WorkoutId.eq(input.workout_id.clone())),
-        )
-        .one(db)
-        .await?
-    {
-        let mut to_update: collection_to_entity::ActiveModel = etc.into();
-        to_update.last_updated_on = ActiveValue::Set(Utc::now());
-        to_update.update(db).await?
-    } else {
-        let created_collection = collection_to_entity::ActiveModel {
-            collection_id: ActiveValue::Set(collection.id),
-            information: ActiveValue::Set(input.information),
-            person_id: ActiveValue::Set(input.person_id.clone()),
-            workout_id: ActiveValue::Set(input.workout_id.clone()),
-            metadata_id: ActiveValue::Set(input.metadata_id.clone()),
-            exercise_id: ActiveValue::Set(input.exercise_id.clone()),
-            metadata_group_id: ActiveValue::Set(input.metadata_group_id.clone()),
-            ..Default::default()
-        };
-        let created = created_collection.insert(db).await?;
-        tracing::debug!("Created collection to entity: {:?}", created);
-        if input.workout_id.is_none() {
-            associate_user_with_entity(
-                user_id,
-                input.metadata_id,
-                input.person_id,
-                input.exercise_id,
-                input.metadata_group_id,
-                db,
-            )
-            .await
-            .ok();
-        }
-        created
-    };
-    perform_core_application_job
-        .enqueue(CoreApplicationJob::EntityAddedToCollection(
-            user_id.to_owned(),
-            resp.id,
-        ))
-        .await
-        .unwrap();
-    Ok(true)
 }
 
 #[derive(MergedObject, Default)]
