+++ apps/backend/src/fitness/resolver.rs
@@ -29,11 +29,9 @@ use services::FileStorageService;
 use slug::slugify;
 use traits::AuthProvider;
 use utils::{
-    entity_in_collections, ilike_sql, user_by_id, user_measurements_list, workout_details,
-    GraphqlRepresentation,
+    add_entity_to_collection, entity_in_collections, ilike_sql, user_by_id, user_measurements_list,
+    workout_details, GraphqlRepresentation,
 };
-
-use crate::app_utils::add_entity_to_collection;
 
 use super::logic::{calculate_and_commit, delete_existing_workout};
