+++ apps/backend/src/miscellaneous.rs
@@ -94,13 +94,11 @@ use serde::{Deserialize, Serialize};
 use services::{send_notification, sign, FileStorageService, IntegrationService};
 use traits::{AuthProvider, MediaProvider, MediaProviderLanguages, TraceOk};
 use utils::{
-    apply_collection_filter, convert_naive_to_utc, entity_in_collections, get_current_date,
-    get_first_and_last_day_of_month, ilike_sql, review_by_id, user_by_id, user_id_from_token,
-    IsFeatureEnabled, AUTHOR, SHOW_SPECIAL_SEASON_NAMES, TEMP_DIR, VERSION,
+    add_entity_to_collection, apply_collection_filter, convert_naive_to_utc, entity_in_collections,
+    get_current_date, get_first_and_last_day_of_month, ilike_sql, review_by_id, user_by_id,
+    user_id_from_token, IsFeatureEnabled, AUTHOR, SHOW_SPECIAL_SEASON_NAMES, TEMP_DIR, VERSION,
 };
 use uuid::Uuid;
-
-use crate::app_utils::add_entity_to_collection;
 
 type Provider = Box<(dyn MediaProvider + Send + Sync)>;
